+++ src/main/java/de/cismet/cids/custom/beans/belis/MaterialLeitungCustomBean.java
@@ -40,7 +40,7 @@ public class MaterialLeitungCustomBean extends BaseEntity implements MaterialLei
 
     //~ Instance fields --------------------------------------------------------
 
-    private Long id;
+    private Integer id;
     private String bezeichnung;
 
     //~ Constructors -----------------------------------------------------------
@@ -73,13 +73,13 @@ public class MaterialLeitungCustomBean extends BaseEntity implements MaterialLei
     }
 
     @Override
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
     @Override
-    public void setId(final Long id) {
-        final Long old = this.id;
+    public void setId(final Integer id) {
+        final Integer old = this.id;
         this.id = id;
         this.propertyChangeSupport.firePropertyChange(PROP__ID, old, this.id);
     }
