+++ src/main/java/de/cismet/cids/custom/beans/belis/MauerlascheCustomBean.java
@@ -71,7 +71,7 @@ public class MauerlascheCustomBean extends GeoBaseEntity implements Mauerlasche 
 
     private Integer id;
     private Date erstellungsjahr;
-    private Short laufende_nummer;
+    private Integer laufende_nummer;
     private GeomCustomBean fk_geom;
     private MaterialMauerlascheCustomBean fk_material;
     private TkeyStrassenschluesselCustomBean fk_strassenschluessel;
@@ -151,7 +151,7 @@ public class MauerlascheCustomBean extends GeoBaseEntity implements Mauerlasche 
      *
      * @return  DOCUMENT ME!
      */
-    public Short getLaufende_nummer() {
+    public Integer getLaufende_nummer() {
         return laufende_nummer;
     }
 
@@ -160,8 +160,8 @@ public class MauerlascheCustomBean extends GeoBaseEntity implements Mauerlasche 
      *
      * @param  laufende_nummer  DOCUMENT ME!
      */
-    public void setLaufende_nummer(final Short laufende_nummer) {
-        final Short old = this.laufende_nummer;
+    public void setLaufende_nummer(final Integer laufende_nummer) {
+        final Integer old = this.laufende_nummer;
         this.laufende_nummer = laufende_nummer;
         this.propertyChangeSupport.firePropertyChange(PROP__LAUFENDE_NUMMER, old, this.laufende_nummer);
     }
@@ -307,12 +307,12 @@ public class MauerlascheCustomBean extends GeoBaseEntity implements Mauerlasche 
     }
 
     @Override
-    public Short getLaufendeNummer() {
+    public Integer getLaufendeNummer() {
         return getLaufende_nummer();
     }
 
     @Override
-    public void setLaufendeNummer(final Short laufendeNummer) {
+    public void setLaufendeNummer(final Integer laufendeNummer) {
         setLaufende_nummer(laufendeNummer);
     }
